refactor: narrow enter frontend api types
