+++ enter.pollinations.ai/frontend/src/tsconfig.json
@@ -11,7 +11,6 @@
     "include": [
         "**/*.ts",
         "**/*.tsx",
-        "../../worker-runtime.d.ts",
-        "../../worker-bindings.d.ts"
+        "../../worker-runtime.d.ts"
     ]
 }
